Split the OPS and ADMS validators
Did use a base class, but that's more to avoid redefining the common errors (and it might be used for validation of the few base properties as well, though I don't know what those would be).
